+++ src/main/java/ch/ethz/eyetap/dto/ImportWordBoundingBoxDto.java
@@ -7,9 +7,9 @@ import ch.ethz.eyetap.model.annotation.CharacterBoundingBox;
  */
 public record ImportWordBoundingBoxDto(Long foreignId,
                                        String word,
-                                       Long xMin,
-                                       Long xMax,
-                                       Long yMin,
-                                       Long yMax
+                                       Double xMin,
+                                       Double xMax,
+                                       Double yMin,
+                                       Double yMax
 ) {
 }
